Moved SpectrumAnalyzer and use direct import
