+++ src/renderer/components/visualizers/SpectrumAnalyzer.vue
