--- src/renderer/components/visualizers/SpectrumAnalyzer/index.ts
@@ -1,5 +1,0 @@
-import SpectrumAnalyzer from "./SpectrumAnalyzer.vue";
-
-export {
-  SpectrumAnalyzer
-};
